common: Fix progress update on UT file downloads

+++ src/common/src/commands/ubuntutouch/utfiledownloadcommandentity.cpp
@@ -96,7 +96,7 @@ bool UtFileDownloadCommandEntity::abortWork()
 
 void UtFileDownloadCommandEntity::handleProgress(qulonglong received, qulonglong total)
 {
-    setProgress((qreal)(received/total));
+    setProgress((qreal)((qreal)received/(qreal)total));
 }
 
 void UtFileDownloadCommandEntity::handleFinished(const QString& path)
